# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qu\u1ea3n l\u00fd Tao API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/collector/registry": {
      "get": {
        "tags": [
          "Collector Ingest"
        ],
        "summary": "Get Collector Registry",
        "operationId": "get_collector_registry_v1_collector_registry_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/collector/events": {
      "post": {
        "tags": [
          "Collector Ingest"
        ],
        "summary": "Ingest Events",
        "operationId": "ingest_events_v1_collector_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/CollectorEventItem"
                },
                "type": "array",
                "title": "Events"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/collector/heartbeat": {
      "post": {
        "tags": [
          "Collector Ingest"
        ],
        "summary": "Collector Heartbeat",
        "operationId": "collector_heartbeat_v1_collector_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/transactions": {
      "post": {
        "tags": [
          "Ledger"
        ],
        "summary": "Create Transaction Endpoint",
        "operationId": "create_transaction_endpoint_v1_transactions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "List Transactions Endpoint",
        "operationId": "list_transactions_endpoint_v1_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(income|expense)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Get Transaction Endpoint",
        "operationId": "get_transaction_endpoint_v1_transactions__transaction_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Ledger"
        ],
        "summary": "Update Transaction Endpoint",
        "operationId": "update_transaction_endpoint_v1_transactions__transaction_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ledger"
        ],
        "summary": "Delete Transaction Endpoint",
        "operationId": "delete_transaction_endpoint_v1_transactions__transaction_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/period": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Get Period Reports Endpoint",
        "operationId": "get_period_reports_endpoint_v1_reports_period_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO datetime start boundary",
              "title": "Start Date"
            },
            "description": "ISO datetime start boundary"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO datetime end boundary",
              "title": "End Date"
            },
            "description": "ISO datetime end boundary"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sync/snapshot": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Get Snapshot Endpoint",
        "operationId": "get_snapshot_endpoint_v1_sync_snapshot_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "known_revision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Known Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sync/operations": {
      "post": {
        "tags": [
          "Sync"
        ],
        "summary": "Post Operations Endpoint",
        "operationId": "post_operations_endpoint_v1_sync_operations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationsBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectorEventItem": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "collector_epoch": {
            "type": "integer",
            "title": "Collector Epoch"
          },
          "binding_id": {
            "type": "string",
            "title": "Binding Id"
          },
          "binding_version": {
            "type": "integer",
            "title": "Binding Version"
          },
          "capture_epoch": {
            "type": "integer",
            "title": "Capture Epoch"
          },
          "bank_code": {
            "type": "string",
            "title": "Bank Code"
          },
          "owner_account": {
            "type": "string",
            "title": "Owner Account"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "amount_vnd": {
            "type": "string",
            "title": "Amount Vnd"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "time_source": {
            "type": "string",
            "title": "Time Source"
          },
          "bank_description": {
            "type": "string",
            "title": "Bank Description"
          },
          "bank_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Reference"
          },
          "source_package": {
            "type": "string",
            "title": "Source Package"
          },
          "parser_version": {
            "type": "string",
            "title": "Parser Version"
          },
          "source_event_key": {
            "type": "string",
            "title": "Source Event Key"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "collector_epoch",
          "binding_id",
          "binding_version",
          "capture_epoch",
          "bank_code",
          "owner_account",
          "direction",
          "amount_vnd",
          "occurred_at",
          "time_source",
          "bank_description",
          "source_package",
          "parser_version",
          "source_event_key"
        ],
        "title": "CollectorEventItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeartbeatRequest": {
        "properties": {
          "queue_size": {
            "type": "integer",
            "title": "Queue Size",
            "default": 0
          },
          "listener_connected": {
            "type": "boolean",
            "title": "Listener Connected",
            "default": true
          },
          "counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counters",
            "default": {}
          }
        },
        "type": "object",
        "title": "HeartbeatRequest"
      },
...
      "OperationsBatchRequest": {
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/TypedOperation"
            },
            "type": "array",
            "title": "Operations"
          }
        },
        "type": "object",
        "required": [
          "operations"
        ],
        "title": "OperationsBatchRequest"
      },
...
      "TransactionCreateRequest": {
        "properties": {
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "amount_vnd": {
            "type": "string",
            "title": "Amount Vnd"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "tag_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tag Ids",
            "default": []
          },
          "user_note": {
            "type": "string",
            "title": "User Note",
            "default": ""
          },
          "purpose": {
            "type": "string",
            "title": "Purpose",
            "default": "normal"
          }
        },
        "type": "object",
        "required": [
          "direction",
          "amount_vnd",
          "occurred_at",
          "category_id"
        ],
        "title": "TransactionCreateRequest"
      },
      "TransactionUpdateRequest": {
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids"
          },
          "user_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Note"
          },
          "amount_vnd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Vnd"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          }
        },
        "type": "object",
        "title": "TransactionUpdateRequest"
      },
      "TypedOperation": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "operation_id",
          "type",
          "payload"
        ],
        "title": "TypedOperation"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}